add handling for converting legacy omage.def files
Changed format of legacy `omega.def` file to be usable by new program.
Putting all the values on one line does not change how the legacy
program would operate since newline and normal whitespace is treated the
same.

Updated `matrix_operations.is_vector()` and related functions so that a
1x1 matrix is considered a vector and square matrix.

Updated loading and writing matrix definition files
- `matrix_operations.load_files()` will now return a dictionary of
  objects that were read from individual files.
- `matrix_operations.write_files()` now only accepts a dictionary of
  objects to write to files and will use the dictionary keys to
  determine the name of the file to write.
- Added handling of reading `omega.def` files in both legacy and new
  format.

Updated instantiation of solvers to expect a dictionary of objects that
contain the items to pass to constructors.

Resolves #50