solver constructors now use **kwargs
Updated solver constructors to accept only `**kwargs` argument instead
of individually named arguments. This makes it easier for classes in the
solver hierarchy to perform some of the work in input validation. It
also helps to simplfy how arguments get passed to the constructors.

Updated tests to suit the new paradigm. Also, mode the logic used for
`test_solve` generic to be used my other solvers.

Resolves #21